[style/#2]:type 분리
